requirements.txt
google-cloud-speech
google-cloud-dialogflow
google-cloud-texttospeech
pyaudio
pydub
playsound 